requirements.txt
bibtexparser==1.3.0
Markdown==3.4.1
pyparsing==3.0.9
PyYAML==6.0
